test: rename `@tour-of-heroes/crisis-center` to `@tour-of-heroes-classic/crisis-center`
