+++ packages/examples/tour-of-heroes-classic/crisis-center/ng-package.json
@@ -1,6 +1,6 @@
 {
   "$schema": "../../../../node_modules/ng-packagr/ng-package.schema.json",
-  "dest": "../../../../dist/packages/examples/tour-of-heroes/crisis-center",
+  "dest": "../../../../dist/packages/examples/tour-of-heroes-classic/crisis-center",
   "lib": {
     "entryFile": "src/index.ts"
   }
