+++ packages/examples/tour-of-heroes-classic/crisis-center/package.json
@@ -1,0 +1,3 @@
+{
+  "name": "@tour-of-heroes-classic/crisis-center"
+}
