+++ packages/examples/tour-of-heroes-classic/crisis-center/project.json
@@ -1,22 +1,22 @@
 {
-  "name": "crisis-center",
+  "name": "examples-tour-of-heroes-classic-crisis-center",
   "$schema": "../../../../node_modules/nx/schemas/project-schema.json",
   "projectType": "library",
-  "sourceRoot": "packages/examples/tour-of-heroes/crisis-center/src",
+  "sourceRoot": "packages/examples/tour-of-heroes-classic/crisis-center/src",
   "prefix": "app-crisis",
   "targets": {
     "build": {
       "executor": "@nrwl/angular:ng-packagr-lite",
       "outputs": ["{workspaceRoot}/dist/{projectRoot}"],
       "options": {
-        "project": "packages/examples/tour-of-heroes/crisis-center/ng-package.json"
+        "project": "packages/examples/tour-of-heroes-classic/crisis-center/ng-package.json"
       },
       "configurations": {
         "development": {
-          "tsConfig": "packages/examples/tour-of-heroes/crisis-center/tsconfig.lib.json"
+          "tsConfig": "packages/examples/tour-of-heroes-classic/crisis-center/tsconfig.lib.json"
         },
         "production": {
-          "tsConfig": "packages/examples/tour-of-heroes/crisis-center/tsconfig.lib.prod.json"
+          "tsConfig": "packages/examples/tour-of-heroes-classic/crisis-center/tsconfig.lib.prod.json"
         }
       },
       "defaultConfiguration": "production"
@@ -25,8 +25,8 @@
       "executor": "@nrwl/linter:eslint",
       "options": {
         "lintFilePatterns": [
-          "packages/examples/tour-of-heroes/crisis-center/src/**/*.ts",
-          "packages/examples/tour-of-heroes/crisis-center/src/**/*.html"
+          "packages/examples/tour-of-heroes-classic/crisis-center/src/**/*.ts",
+          "packages/examples/tour-of-heroes-classic/crisis-center/src/**/*.html"
         ]
       }
     },
@@ -34,7 +34,7 @@
       "executor": "@nrwl/jest:jest",
       "outputs": ["{workspaceRoot}/coverage/{projectRoot}"],
       "options": {
-        "jestConfig": "packages/examples/tour-of-heroes/crisis-center/jest.config.ts",
+        "jestConfig": "packages/examples/tour-of-heroes-classic/crisis-center/jest.config.ts",
         "passWithNoTests": true
       }
     }
