+++ packages/examples/tour-of-heroes-classic/crisis-center/src/lib/can-deactivate.guard.ts
@@ -10,7 +10,8 @@ export interface CanComponentDeactivate {
   providedIn: 'root',
 })
 export class CanDeactivateGuard
-  implements CanDeactivate<CanComponentDeactivate> {
+  implements CanDeactivate<CanComponentDeactivate>
+{
   canDeactivate(component: CanComponentDeactivate) {
     return component.canDeactivate ? component.canDeactivate() : true;
   }
