+++ packages/examples/tour-of-heroes-classic/crisis-center/src/lib/crisis-center.module.ts
@@ -1,11 +1,11 @@
+import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { FormsModule } from '@angular/forms';
-import { CommonModule } from '@angular/common';
 
 import { CrisisCenterHomeComponent } from './crisis-center-home/crisis-center-home.component';
-import { CrisisListComponent } from './crisis-list/crisis-list.component';
 import { CrisisCenterComponent } from './crisis-center/crisis-center.component';
 import { CrisisDetailComponent } from './crisis-detail/crisis-detail.component';
+import { CrisisListComponent } from './crisis-list/crisis-list.component';
 
 import { CrisisCenterRoutingModule } from './crisis-center-routing.module';
