+++ packages/examples/tour-of-heroes-classic/crisis-center/src/lib/crisis-detail-resolver.service.ts
