+++ packages/examples/tour-of-heroes-classic/crisis-center/src/lib/crisis-detail/crisis-detail.component.html
