+++ packages/examples/tour-of-heroes-classic/crisis-center/src/lib/crisis-list/crisis-list.component.css
