+++ packages/examples/tour-of-heroes-classic/crisis-center/src/lib/fake-dialog.service.ts
