+++ packages/examples/tour-of-heroes-classic/crisis-center/src/lib/mock-crises.ts
