+++ packages/examples/tour-of-heroes-classic/crisis-center/tsconfig.lib.prod.json
