+++ packages/examples/tour-of-heroes-classic/crisis-center/tsconfig.spec.json
