--- packages/examples/tour-of-heroes/crisis-center/package.json
@@ -1,3 +1,0 @@
-{
-  "name": "@tour-of-heroes/crisis-center"
-}
