+++ packages/spectacular/src/lib/feature-testing/integration/angular-testing-library/crisis-center.feature.spec.ts
@@ -7,7 +7,7 @@ import {
   CrisisService,
   DialogService,
   FakeDialogService,
-} from '@tour-of-heroes/crisis-center';
+} from '@tour-of-heroes-classic/crisis-center';
 import { SpectacularAppComponent } from '../../../shared/app-component/spectacular-app.component';
 import { provideSpectacularFeatureTesting } from '../../configuration/provide-spectacular-feature-testing';
 import { SpectacularFeatureLocation } from '../../navigation/spectacular-feature-location';
@@ -45,7 +45,10 @@ async function setup() {
       provideSpectacularFeatureTesting({
         featurePath: crisisCenterPath,
         routes: [
-          { path: crisisCenterPath, loadChildren: () => CrisisCenterModule },
+          {
+            path: crisisCenterPath,
+            loadChildren: () => CrisisCenterModule,
+          },
         ],
       }),
       { provide: DialogService, useClass: FakeDialogService },
